msm: kgsl: Wake GPU upon receiving an ioctl rather than upon touch input
Waking the GPU upon touch wastes power when the screen is being touched
in a way that does not induce animation or any actual need for GPU usage.
Instead of preemptively waking the GPU on touch input, wake it up upon
receiving a IOCTL_KGSL_GPU_COMMAND ioctl since it is a sign that the GPU
will soon be needed.

Change-Id: I4c1525915930da224388ff3a3b2f8b3d303f79c9
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>